added extra in decl and prevented symlink at non-existing files

+++ install.sh
@@ -1,11 +1,13 @@
 #! /bin/bash
 
-declare -a files=(aliases bash_profile bash_prompt exports functions gitconfig ackrc)
+declare -a files=(aliases bash_profile bash_prompt exports functions gitconfig ackrc extra)
 
 # loop through the files
 for file in ${files[@]}; do
     sourceFile="$(pwd)/$file"
     targetFile="$HOME/$(printf "%s" ".$file" | sed "s/.*\/\(.*\)/\1/g")"
-
-    ln -fs $sourceFile $targetFile
+    
+    if [ -e $file ]; then
+        ln -fs $sourceFile $targetFile
+    fi
 done
